chore: icons 폴더 변경 및 package.json script 수정
