+++ packages/design-system/package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "lint": "eslint --max-warnings 0",
     "generate:component": "turbo gen react-component",
-    "svgr": "svgr --out-dir src/icons/src/svg src/icons/assets --typescript --icon && mv src/icons/src/svg/index.ts src/icons/src/index.ts",
+    "svgr": "svgr --out-dir src/icons/src src/icons/assets --typescript --icon",
     "optimize-svg": "pnpm exec svgo -f src/icons/assets -o src/icons/assets",
     "build": "npm run optimize-svg",
     "storybook": "storybook dev -p 6006",
