--- packages/design-system/src/icons/src/svg/SymbolLogo.tsx
@@ -1,19 +1,0 @@
-import type { SVGProps } from 'react';
-const SvgSymbolLogo = (props: SVGProps<SVGSVGElement>) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="1em"
-    height="1em"
-    fill="none"
-    viewBox="0 0 48 49"
-    {...props}
-  >
-    <path
-      fill="#000"
-      fillRule="evenodd"
-      d="M36.438 3.187 29.74 9.97V.525H17.682v9.444l-6.484-6.564-8.535 8.623 6.302 6.38H0v5.576h18.112c1.158 0 2.315-.285 3.225-1.022 2.3-1.841 2.696-5.391 2.696-12.206 0 6.815.397 10.365 2.696 12.206.91.72 2.084 1.022 3.226 1.022h18.029v-5.425l-.15-.151H38.44l6.517-6.597-8.535-8.64zM48 25.524H30.352a5.6 5.6 0 0 0-2.283.47c-3.49 1.59-4.02 5.006-4.02 12.96 0-5.057-.727-8.507-1.918-10.667-.976-1.758-2.878-2.763-4.88-2.763H0v5.107h9.329l-6.633 6.715 8.535 8.64 6.45-6.53V49H29.74v-9.544l6.848 6.932h.033l8.502-8.623-7.03-7.117H48v-5.124"
-      clipRule="evenodd"
-    />
-  </svg>
-);
-export default SvgSymbolLogo;
